feat(webhook): reject MaxUnavailable greater than a component's replicas
Add a PodCliqueSet validation rejecting a component MaxUnavailable that exceeds
its replica count. A standalone PodClique is checked against its own replicas and
a PodCliqueScalingGroup against its configured replicas, so a rollout cannot be
configured to disrupt more units than exist.

Also move defaultUpdateStrategy into defaultPodCliqueSetSpec so spec-level
defaulting mirrors the type tree, and merge the defaulting tests into a single
podcliqueset_test.go.
